重构公开 API 与通用类型，并支持点击定位无人机。
收拢协议细节与内部配置入口到 Private/Common，统一文案与遥测等配置；选中无人机时地图定位到其 GPS。

+++ Src/Common/QGpsPosition.cpp
@@ -1,14 +1,13 @@
 #include <QtGlobal>
-#include "AirLine/QGpsPosition.h"
+#include "Common/QGpsPosition.h"
 
-// QGpsPosition 实现
 QGpsPosition::QGpsPosition()
-    : m_longitude(0.0), m_latitude(0.0), m_altitude(0.0)
+    : m_longitude(0.0), m_latitude(0.0), m_altitude(0.0f)
 {
 }
 
 QGpsPosition::QGpsPosition(double longitude, double latitude, double altitude)
-    : m_longitude(longitude), m_latitude(latitude), m_altitude(altitude)
+    : m_longitude(longitude), m_latitude(latitude), m_altitude(static_cast<float>(altitude))
 {
 }
 
@@ -38,4 +37,3 @@ bool QGpsPosition::operator!=(const QGpsPosition &other) const
 {
     return !(*this == other);
 }
-
